+++ macos/file_saver/Sources/file_saver/FileSaverPlugin.swift
@@ -45,6 +45,8 @@ struct Params {
     let sourcePath: String?
     let fileExtension: String?
     let includeExtension: Bool
+    let initialDirectory: String?
+    let dialogTitle: String?
     init(_ d: [String: Any?]) {
         fileName = d["name"] as? String
         let uint8List = d["bytes"] as? FlutterStandardTypedData
@@ -56,5 +58,7 @@ struct Params {
         sourcePath = d["sourcePath"] as? String
         fileExtension = d["fileExtension"] as? String
         includeExtension = d["includeExtension"] as? Bool ?? true
+        initialDirectory = d["initialDirectory"] as? String
+        dialogTitle = d["dialogTitle"] as? String
     }
 }